Combining into one generator
